% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{penalty}
\alias{penalty}
\title{Penalty function on the composite L1, L2, and ridge penalty}
\usage{
penalty(x, lambda, hierarchy, d)
}
\arguments{
\item{x}{A numeric vector of two.}

\item{lambda}{a vector of three penalty parameters. \eqn{\lambda[1]}
is the L2 penalty for x, \eqn{\lambda[2]} is the ridge penalty
for x, and \eqn{\lambda[3]} is the L1 penalty for x[2], respectively.}

\item{hierarchy}{Indicator variable for 0, 1, 2. 0 is for no overlap, 1 for 
composite L1 and L2 penalty, and 2 for composite L1, L2 and ridge 
penalty, respectively.}

\item{d}{indices for overlapped variables in x.}
}
\description{
Penalty function on the composite L1, L2, and ridge penalty
}
